+++ src/Active.php
@@ -2,6 +2,11 @@
 
 namespace Juy\ActiveMenu;
 
+/**
+ * Class Active
+ * 
+ * @package Juy\Providers
+ */
 class Active
 {
     /**
@@ -22,7 +27,7 @@ class Active
     }
 
     /**
-     * Active class
+     * Active route name
      * If route matches given route (or array of routes) return active class
      *
      * @param $routePattern